+++ packages/api/cms-api/src/access-log/access-log.interceptor.ts
@@ -4,8 +4,8 @@ import { GraphQLResolveInfo } from "graphql";
 import { getClientIp } from "request-ip";
 
 import { CurrentUser } from "../user-permissions/dto/current-user";
-import { SHOULD_LOG_REQUEST } from "./access-log.constants";
-import { ShouldLogRequest } from "./access-log.module";
+import { ACCESS_LOG_CONFIG } from "./access-log.constants";
+import { AccessLogConfig } from "./access-log.module";
 
 const IGNORED_PATHS = ["/dam/images/:hash/:fileId", "/dam/files/:hash/:fileId", "/dam/images/preview/:fileId", "/dam/files/preview/:fileId"];
 
@@ -13,7 +13,7 @@ const IGNORED_PATHS = ["/dam/images/:hash/:fileId", "/dam/files/:hash/:fileId", 
 export class AccessLogInterceptor implements NestInterceptor {
     protected readonly logger = new Logger(AccessLogInterceptor.name);
 
-    constructor(@Optional() @Inject(SHOULD_LOG_REQUEST) private readonly shouldLogRequest?: ShouldLogRequest) {}
+    constructor(@Optional() @Inject(ACCESS_LOG_CONFIG) private readonly config?: AccessLogConfig) {}
 
     intercept(context: ExecutionContext, next: CallHandler) {
         const requestType = context.getType().toString();
@@ -24,10 +24,16 @@ export class AccessLogInterceptor implements NestInterceptor {
         if (requestType === "graphql") {
             const graphqlExecutionContext = GqlExecutionContext.create(context);
             const graphqlContext = graphqlExecutionContext.getContext();
+            const gqlInfo = graphqlExecutionContext.getInfo<GraphQLResolveInfo>();
+
+            if (this.isResolvingGraphQLField(gqlInfo)) {
+                return next.handle();
+            }
 
             if (
-                this.shouldLogRequest &&
-                !this.shouldLogRequest({
+                this.config &&
+                this.config.shouldLogRequest &&
+                !this.config.shouldLogRequest({
                     user: graphqlContext.req.user,
                     req: graphqlContext.req,
                 })
@@ -40,7 +46,6 @@ export class AccessLogInterceptor implements NestInterceptor {
             this.pushUserToRequestData(graphqlContext.req.user, requestData);
 
             const gqlArgs = { ...graphqlExecutionContext.getArgs() };
-            const gqlInfo = graphqlExecutionContext.getInfo<GraphQLResolveInfo>();
 
             if (gqlInfo.operation.operation === "mutation") {
                 delete gqlArgs["input"];
@@ -57,8 +62,9 @@ export class AccessLogInterceptor implements NestInterceptor {
 
             if (
                 IGNORED_PATHS.some((ignoredPath) => httpRequest.route.path.includes(ignoredPath)) ||
-                (this.shouldLogRequest &&
-                    !this.shouldLogRequest({
+                (this.config &&
+                    this.config.shouldLogRequest &&
+                    !this.config.shouldLogRequest({
                         user: httpRequest.user,
                         req: httpRequest,
                     }))
@@ -84,7 +90,12 @@ export class AccessLogInterceptor implements NestInterceptor {
 
     private pushUserToRequestData(user: CurrentUser, requestData: string[]) {
         if (user) {
-            requestData.push(`user: ${user.id}`);
+            requestData.push(this.config && this.config.userToLog ? this.config.userToLog(user) : `user: ${user.id}`);
         }
+    }
+
+    private isResolvingGraphQLField(gqlInfo: GraphQLResolveInfo): boolean {
+        const parentType = gqlInfo.parentType.name;
+        return parentType !== "Query" && parentType !== "Mutation";
     }
 }