+++ packages/admin/admin/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @comet/admin
 
+## 6.8.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.8.0
+
+## 6.7.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.7.0
+
+## 6.6.2
+
+### Patch Changes
+
+-   @comet/admin-icons@6.6.2
+
+## 6.6.1
+
+### Patch Changes
+
+-   @comet/admin-icons@6.6.1
+
+## 6.6.0
+
+### Minor Changes
+
+-   95b97d768: useDataGridRemote: Add `initialFilter` option
+
+    **Example usage:**
+
+    ```tsx
+    const dataGridProps = useDataGridRemote({
+        initialFilter: { items: [{ columnField: "description", operatorValue: "contains", value: "text" }] },
+    });
+    ```
+
+### Patch Changes
+
+-   6b04ac9a4: Fix the key for accessing the themes `styleOverrides` and `defaultProps` of `CometAdminMenu`
+    -   @comet/admin-icons@6.6.0
+
 ## 6.5.0
 
 ### Minor Changes